Let the caller handle send errors
For the moment, ignore any send errors in both directions. Error messages
are printed to the shell if there is an error (badmatch).

Likely the caller wants the tunnel to stay up for transient issues, for
example if the interface goes down ({error,enetunreach}) and might come
up again. In the case of fatal issues (bad fd, ...), we should shutdown
the tunnel.

There are 2 problems:

* the message flow will not be in one direction, e.g.,

    socket -> filter -> tun

  Instead the filter would call into the socket/tun and shut down the
  gen_server killing the tunnel. This makes the system hard to
  understand.

* tuncer:send/2 is a wrapper around erlang:port_command/2. Not sure what
  values it returns. The documentation says badarg. tuncer wraps the
  port_command in a try/catch and returns an error tuple.

  gen_udp:send returns {error, posix()}. tuncer should do the same.